requirements.txt
eventlet
python-socketio
pillow
qrcode[pil]